fix: rename text area component
